+++ xenon-common/src/test/java/com/vmware/xenon/common/test/TestProperty.java
@@ -29,5 +29,6 @@ public enum TestProperty {
     SET_CONTEXT_ID,
     CALLBACK_SEND,
     PERSISTED,
+    BINARY_SERIALIZATION,
     HTTP2
 }